splitted sound code into several classes, need tests
