+++ src/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory(sdl)
 add_subdirectory(utils)
+add_subdirectory(waveform)
