+++ src/sdl/CMakeLists.txt
@@ -34,4 +34,4 @@ target_link_libraries(
 )
 
 target_link_libraries(${TARGET} PRIVATE spdlog::spdlog)
-target_link_libraries(${TARGET} PRIVATE tools-utils)
+target_link_libraries(${TARGET} PRIVATE tools-utils tools-waveform)
